updated dbengine unittest (#17232)
* when a page cannot be acquired, repeat the call until it can or does not exist

* updated dbengine unittest

* Update dbengine-unittest.c

(cherry picked from commit 4355e50bceb7c6aab409cb671c04577a31c6a868)
